# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.1.0",
    "title": "Road To Ninja Application API",
    "description": "Documentation Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/",
  "tags": [
    {
      "name": "Users",
      "description": "API for users in the system"
    },
    {
      "name": "Projects",
      "description": "API for projects"
    },
    {
      "name": "Authentication",
      "description": "API for authentication"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all users in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Get token",
        "tags": [
          "Authentication"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "credentials",
            "in": "body",
            "required": true,
            "description": "credentials",
            "schema": {
              "$ref": "#/definitions/Credentials"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/TokenInfo"
            }
          }
        }
      }
    },
    "/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get all projects in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Projects"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/project/{id}": {
      "parameters": [
        {
          "name": "name",
          "in": "path",
          "required": true,
          "description": "project id",
          "type": "string"
        }
      ],
      "get": {
        "tags": ["Projects"],
        "summary": "Get project info",
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/Projects"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "x-access-token",
      "in": "header"
    }
  },
  "definitions": {
    "User": {
      "required": [
      ],
      "properties": {
        "username": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "admin": {
          "type": "boolean"
        }
      }
    },
    "Project":{
      "required": [
        "title",
        "uid"
      ],
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "data": {
          "type": "string"
        },
        "requires" : {
          "type": "array"
        },
        "uid": {
          "type": "string"
        }
      },
      "example": {
        "title": "",
        "description": "",
        "data": "",
        "requires" : "",
        "uid": ""
      }
    },
    "Credentials": {
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
         "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "TokenInfo": {
      "properties": {
        "token": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "logged": {
          "type": "boolean"
        }
      }
    },
    "Users": {
      "type": "array",
      "$ref": "#/definitions/User"
    },
    "Projects": {
      "type": "array",
      "$ref": "#/definitions/Project"
    }
  }
}